elisa/elisa-server/requirements.txt
Django==2.2.1
django-cors-headers==2.5.1
django-filter==2.1.0
django-import-export==1.1.0
django-mptt==0.9.0
django-rest-swagger==2.1.2
djangorestframework-simplejwt==4.1.0
djangorestframework==3.9.1
drf-extensions==0.4.0
psycopg2==2.7.7
django-tenants==2.2.3
django-python3-ldap==0.11.2
django-fsm==2.6.1
cx-Oracle==7.1.3
python-decouple==3.1